funciones buscarPorID, listado en modo tabla

+++ src/flujos/AdminMenu.java
@@ -41,7 +41,7 @@ public class AdminMenu {
                     case 4: eliminarAdministrador(sc);break;
                     case 5: iniciarSesion(sc);break;
                     case 6: {return;}
-                    default:System.out.printf("Opcion no valido");
+                    default:System.out.printf("Opción no válida");
                 }
             }while (op != 6);
         }catch (SQLException ex){
@@ -61,10 +61,10 @@ public class AdminMenu {
 
     public void crearAdministrador(Scanner sc) throws SQLException {
 
-        System.out.println("Ingrese nombre del  administrador: ");
+        System.out.println("Ingrese nombre del administrador: ");
         String nombre = sc.nextLine();
 
-        System.out.println("Ingrese fecha Nacimiento del administrador(formato: yyyy-mm-dd): ");
+        System.out.println("Ingrese fecha Nacimiento del administrador(formato: YYYY-MM-DD): ");
         Date fechaNacimiento = Date.valueOf(sc.next());
 
         sc.nextLine();
@@ -85,6 +85,12 @@ public class AdminMenu {
         int id = sc.nextInt();
         sc.nextLine();
 
+        Administrador admin = admindao.buscarAdminPorId(id);
+        if(admin == null) {
+            System.out.println("El administrador no existe");
+            return;
+        }
+
         admindao.eliminarAdministrador(id);
     }
 
@@ -94,6 +100,12 @@ public class AdminMenu {
         System.out.println("Ingrese el id del administrador a modificar: ");
         int id = sc.nextInt();
         sc.nextLine();
+
+        Administrador admin = admindao.buscarAdminPorId(id);
+        if(admin == null) {
+            System.out.println("El administrador no existe");
+            return;
+        }
 
         System.out.println("Ingrese nuevo nombre del administrador: ");
         String nombre = sc.nextLine();
@@ -113,6 +125,8 @@ public class AdminMenu {
         if (administrador.isEmpty()) {
             System.out.println("No hay administradores para mostrar");
         } else {
+            System.out.printf("%-5s %-10s %-30s %-30s %-20s%n",
+                    "ID","Nombre","Fecha Nacimiento","Correo","Contraseña");
             for (Administrador administrador1 : administrador) {
                 administrador1.mostrarInfo();
             }
